Checking in
Some pretty big refactors. Works worse so far but code is cleaner

+++ core/src/main/java/com/internal/bootstrap/geometrypipeline/dynamicgeometrymanager/dynamicgeometry/PartialGeometryBranch.java
@@ -23,7 +23,7 @@ public class PartialGeometryBranch extends BranchPackage {
             BlockPaletteHandle biomePaletteHandle,
             BlockPaletteHandle blockPaletteHandle,
             DynamicPacketInstance dynamicPacketInstance,
-            short xyz,
+            int xyz,
             Direction3Vector direction3Vector,
             BiomeHandle biomeHandle,
             BlockHandle blockHandle,
